add hagg data gen, minor change

+++ parquet-benchmarks/src/main/java/org/apache/parquet/benchmarks/DataGenerator.java
@@ -126,7 +126,7 @@ public class DataGenerator {
     for (int i = 0; i < nRows; i++) {
       writer.write(
         f.newGroup()
-          .append("row_count", i)
+          .append("row_count", (long) i)
           .append("gby_int32", i % nGroups)
           .append("gby_string", gbyString.substring(0,i % nGroups))
           .append("gby_float", fl + (i % nGroups))
